+++ tests/dummy/app/controllers/ember-flexberry-dummy-multi-list-user-edit/new.js
@@ -1,0 +1,1 @@
+export { default } from '../ember-flexberry-dummy-multi-list-user-edit';