+++ CHANGELOG.md
@@ -3,13 +3,22 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [2.2.0-beta.0] - 2019-06-28
 ### Added
 * `flexberry-lookup` component:
     * Add select first autocomplete result by enter click.
     * Add dropdown class to flexberry-lookup at drodpdown mode.
-    * Add limitFunction in groupedit row to lookup limit by other component. 
+    * Add limitFunction in groupedit row to lookup limit by other component.
 * `flexberry-file` component:
     * Added the ability correct settings to file preview modal dialog.
+* Add mixins for multi list mode.
+
+### Changed
+* Update dependency on `ember-flexberry-data` to version `2.2.0-beta.0`.
+
+### Breaking changes
+* Removed `flexberry-objectlistview-on-edit-form-controller` mixin from `edit-form` route. Use new mixins or include old mixin in yours forms with olv on edit form.
 
 ## [2.1.0] - 2019-04-30
 ### Added